pyproject.toml
[project]
name = "morpheus-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "attrs>=25.4.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dateutil>=2.9.0.post0",
]
